fix permissionlevels import
