+++ maps/pirates/permissions.lua
@@ -10,6 +10,25 @@ local CoreData = require 'maps.pirates.coredata'
 
 local Public = {}
 
+local privilege_levels = {
+	NORMAL = 1,
+	OFFICER = 2,
+	CAPTAIN = 3
+}
+Public.privilege_levels = privilege_levels
+
+function Public.player_privilege_level(player)
+	local memory = Memory.get_crew_memory()
+
+	if Common.is_id_valid(memory.id) and Common.is_captain(player) then
+		return Public.privilege_levels.CAPTAIN
+	elseif Common.is_officer(player.index) then
+		return Public.privilege_levels.OFFICER
+	else
+		return Public.privilege_levels.NORMAL
+	end
+end
+
 local function set_normal_permissions(group)
 	if not _DEBUG then
 		group.set_allows_action(defines.input_action.edit_permission_group, false)
