+++ maps/pirates/shop/shop.lua
@@ -10,14 +10,14 @@ local Crew = require 'maps.pirates.crew'
 local Balance = require 'maps.pirates.balance'
 local Common = require 'maps.pirates.common'
 -- local Utils = require 'maps.pirates.utils_local'
-local Roles = require 'maps.pirates.roles.roles'
+-- local Roles = require 'maps.pirates.roles.roles'
 local Math = require 'maps.pirates.math'
 local _inspect = require 'utils.inspect'.inspect
 local SurfacesCommon = require 'maps.pirates.surfaces.common'
 local Upgrades = require 'maps.pirates.shop.boat_upgrades'
 local Cabin = require 'maps.pirates.surfaces.cabin'
 -- local Upgrades = require 'maps.pirates.shop.boat_upgrades'
-
+local Permissions = require 'maps.pirates.permissions'
 local Public = {}
 Public.Covered = require 'maps.pirates.shop.covered'
 Public.Merchants = require 'maps.pirates.shop.merchants'
@@ -91,7 +91,7 @@ local function purchaseData(market, player, offer_index)
 	local in_captains_cabin = type and type == SurfacesCommon.enum.CABIN
 	local dock_upgrades_market = destination.dynamic_data.dock_captains_market and (destination.dynamic_data.dock_captains_market == market)
 
-	local permission_level_fail = (in_captains_cabin and Roles.player_privilege_level(player) < Roles.privilege_levels.OFFICER) or (dock_upgrades_market and Roles.player_privilege_level(player) < Roles.privilege_levels.OFFICER)
+	local permission_level_fail = (in_captains_cabin and Permissions.player_privilege_level(player) < Permissions.privilege_levels.OFFICER) or (dock_upgrades_market and Permissions.player_privilege_level(player) < Permissions.privilege_levels.OFFICER)
 
 	if in_captains_cabin then
 		decay_type = 'static'
